requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
openai==1.3.7
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
antlr4-python3-runtime==4.13.1
semantic-kernel==0.4.0.dev0
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1 